HANDLERS: Split in different modules (rundeck, hostaliases, metadata).
