+++ bin/dbod-api
@@ -14,13 +14,16 @@ DB On Demand metadata REST API server
 
 import ConfigParser
 import sys, traceback
+import tornado.web
 
 from tornado.options import parse_command_line, options
 from tornado.httpserver import HTTPServer
 from tornado.ioloop import IOLoop
 
 from dbod.api.base import DocHandler
-from dbod.api.rundeck import *
+from dbod.api.rundeck import RundeckResources
+from dbod.api.metadata import Metadata
+from dbod.api.hostaliases import HostAliases
 from dbod.config import config
 
 import logging
