+++ dbod/api/rundeck.py
@@ -13,9 +13,6 @@ REST API Server for the DB On Demand System
 """
 
 import tornado.web
-import tornado.log
-import base64
-import functools
 import logging
 import json
 import requests
@@ -56,52 +53,3 @@ class RundeckResources(tornado.web.RequestHandler):
                 raise tornado.web.HTTPError(NOT_FOUND)
         else:
             logging.error("Internal Rundeck resources endpoint not configured")
-            
-
-class HostAliases(tornado.web.RequestHandler):
-    def get(self, host):
-        """list of ip-aliases registered in a host"""
-        url = config.get('postgrest', 'host_aliases_url')
-        if url:
-            composed_url = url + '?host=eq.' + host
-            logging.debug('Requesting ' + composed_url )
-            response = requests.get(composed_url)
-            if response.ok:
-                data = json.loads(response.text)
-                d = data.pop()
-                self.write(d.get('aliases'))
-            else: 
-                logging.error("Error fetching aliases in host: " + host)
-                raise tornado.web.HTTPError(NOT_FOUND)
-        else:
-            logging.error("Internal host aliases endpoint not configured")
-            
-class Metadata(tornado.web.RequestHandler):
-    def get(self, **args):
-        """Returns entity metadata"""
-        url = config.get('postgrest', 'entity_metadata_url')
-        name = args.get('name')
-        etype = args.get('class')
-        if url:
-            if etype == u'entity':
-                composed_url = url + '?db_name=eq.' + name
-            else:
-                composed_url = url + '?host=eq.' + name
-            logging.debug('Requesting ' + composed_url )
-            response = requests.get(composed_url)
-            if response.ok:
-                data = json.loads(response.text)
-                if data != []:
-                    if etype == u'entity':
-                        d = data.pop()
-                        self.write(d)
-                    else:
-                        self.write(json.dumps(data))
-                else: 
-                    logging.error("Entity metadata not found: " + name)
-                    raise tornado.web.HTTPError(NOT_FOUND)
-            else: 
-                logging.error("Error fetching entity metadata: " + name)
-                raise tornado.web.HTTPError(response.status_code)
-        else:
-            logging.error("Internal entity metadata endpoint not configured")
